+++ docs/source/examples/plot_ukf_demo.py
@@ -163,6 +163,8 @@ tx_frequency = 1.5 * 1e9  # Sensor transmit center frequency (Hz)
 
 # Exemplar is akin to sensor capability descriptions found in OV-1s (baseball at LEO, basketball at GEO)
 exemplar = [0.04908738521234052, 40500.0]  # Exemplar area (m^2)  # Exemplar range (km)
+field_of_view = "conic"
+calc_field_of_view = True
 
 radar_sensor = Radar(
     az_mask,
@@ -174,6 +176,8 @@ radar_sensor = Radar(
     tx_power,
     tx_frequency,
     np.radians(slew_rate),
+    field_of_view,
+    calc_field_of_view,
 )
 
 # Lat, Lon, Alt for sensor near VT