+++ src/resonaate/scenario/scenario_builder.py
@@ -18,10 +18,11 @@ from ..data.events import Event
 from ..data.resonaate_database import ResonaateDatabase
 from ..dynamics import spacecraftDynamicsFactory
 from ..physics.noise import noiseCovarianceFactory
-from ..scenario.clock import ScenarioClock
 from ..tasking.decisions import decisionFactory
 from ..tasking.engine.centralized_engine import CentralizedTaskingEngine
 from ..tasking.rewards import rewardsFactory
+from .clock import ScenarioClock
+from .config.base import NO_SETTING
 from .config.event_configs import MissingDataDependency
 
 
@@ -143,6 +144,9 @@ class ScenarioBuilder:
 
         self.sensor_network = []
         for agent in sensor_configs.values():
+            # Assign Sensor FoV from init if not set
+            if agent.calculate_fov is NO_SETTING:
+                agent._calculate_fov._setting = self.observation.field_of_view
             config = {
                 "agent": agent,
                 "clock": self.clock,
@@ -299,3 +303,8 @@ class ScenarioBuilder:
     def time_step(self):
         """TimeConfig: returns "time_step" section of the configuration."""
         return self.time.physics_step_sec
+
+    @property
+    def observation(self):
+        """ObservationConfig: returns "observation" section of the configuration."""
+        return self._config.observation