+++ src/resonaate/sensors/sensor_base.py
@@ -1,10 +1,13 @@
 """Abstract :class:`.Sensor` base class which defines the common sensor interface."""
+from __future__ import annotations
+
 # Standard Library Imports
 from abc import ABCMeta, abstractmethod
 from collections import namedtuple
+from typing import TYPE_CHECKING
 
 # Third Party Imports
-from numpy import arccos, asarray, cos, diagflat, dot, fix, matmul, random, real, sin
+from numpy import array, cos, diagflat, dot, matmul, random, real, sin, squeeze
 from scipy.linalg import norm, sqrtm
 
 # Local Imports
@@ -12,9 +15,21 @@ from ..common.logger import resonaateLogInfo
 from ..common.utilities import getTypeString
 from ..data.observation import Observation
 from ..physics import constants as const
+from ..physics.math import safeArccos, subtendedAngle
 from ..physics.sensor_utils import lineOfSight
+from ..physics.time.stardate import ScenarioTime
 from ..physics.transforms.methods import getSlantRangeVector
 from .measurements import getAzimuth, getElevation, getRange
+
+if TYPE_CHECKING:
+    # Third Party Imports
+    from numpy import ndarray
+
+    # Local Imports
+    from ..agents.estimate_agent import EstimateAgent
+    from ..agents.target_agent import TargetAgent
+    from . import FieldOfView
+
 
 ObservationTuple = namedtuple("ObservationTuple", ["observation", "agent", "angles"])
 """Named tuple for correlating an observation, the sensing agent, and its angular values.
@@ -22,23 +37,38 @@ ObservationTuple = namedtuple("ObservationTuple", ["observation", "agent", "angl
 Attributes:
     observation (:class:`.Observation` | ``None``): valid or invalid observation
     agent (:class:`.SensingAgent`): reference to `~.Sensor.host`
-    angles (``np.ndarray``): array signifying which measurements are angles via :class:`.IsAngle` enum.
+    angles (``ndarray``): array signifying which measurements are angles via :class:`.IsAngle` enum.
 """
 
 
 class Sensor(metaclass=ABCMeta):
     """Abstract base class for a generic Sensor object."""
 
-    def __init__(self, az_mask, el_mask, r_matrix, diameter, efficiency, slew_rate, **sensor_args):
+    def __init__(
+        self,
+        az_mask: ndarray,
+        el_mask: ndarray,
+        r_matrix: ndarray,
+        diameter: float,
+        efficiency: float,
+        exemplar: ndarray,
+        slew_rate: float,
+        field_of_view: FieldOfView,
+        calculate_fov: bool,
+        **sensor_args,
+    ):
         """Construct a generic `Sensor` object.
 
         Args:
-            az_mask (``list``): azimuth mask for visibility conditions
-            el_mask (``list``): elevation mask for visibility conditions
-            r_matrix (``np.ndarray``): measurement noise covariance matrix
+            az_mask (``ndarray``): azimuth mask for visibility conditions
+            el_mask (``ndarray``): elevation mask for visibility conditions
+            r_matrix (``ndarray``): measurement noise covariance matrix
             diameter (``float``): diameter of sensor dish (m)
             efficiency (``float``): efficiency percentage of the sensor
+            exemplar (``ndarray``): 2x1 array of exemplar capabilities, used in min detectable power  calculation [cross sectional area (m^2), range (km)]
             slew_rate (``float``): maximum rotational speed of the sensor (deg/sec)
+            field_of_view (:class:`.FieldOfView`): field of view of sensor
+            calculate_fov (``bool``): whether or not to calculate Field of View, default=True
             sensor_args (``dict``): extra key word arguments for easy extension of the `Sensor` interface
         """
         self._az_mask = None
@@ -49,11 +79,13 @@ class Sensor(metaclass=ABCMeta):
         self.r_matrix = r_matrix
         self.aperture_area = const.PI * ((diameter / 2.0) ** 2)
         self.efficiency = efficiency
+        self.exemplar = squeeze(exemplar)
         self.slew_rate = const.DEG2RAD * slew_rate
+        self.field_of_view = field_of_view
+        self.calculate_field_of_view = calculate_fov
 
         # Derived properties initialization
-        self.exemplar = None
-        self.time_last_ob = 0.0
+        self.time_last_ob = ScenarioTime(0.0)
         self.delta_boresight = 0.0
         self._host = None
         self.boresight = None
@@ -64,12 +96,12 @@ class Sensor(metaclass=ABCMeta):
         self._sensor_args = sensor_args
 
     @abstractmethod
-    def maximumRangeTo(self, viz_cross_section, tgt_eci_state):
+    def maximumRangeTo(self, viz_cross_section: float, tgt_eci_state: ndarray) -> float:
         """Calculate the maximum possible range based on a target's visible area.
 
         Args:
             viz_cross_section (``float``): area of the target facing the sun (m^2)
-            tgt_eci_state (``np.ndarray``): 6x1 ECI state vector of the target agent
+            tgt_eci_state (``ndarray``): 6x1 ECI state vector of the target agent
 
         Returns:
             ``float``: maximum possible range to target at which this sensor can make valid observations (km)
@@ -77,15 +109,20 @@ class Sensor(metaclass=ABCMeta):
         raise NotImplementedError
 
     @abstractmethod
-    def getMeasurements(self, slant_range_sez, noisy=False):
+    def getMeasurements(self, slant_range_sez: ndarray, noisy=False) -> dict:
         """Return the measurement state of the measurement.
 
         Args:
-            slant_range_sez (``np.ndarray``): 6x1 SEZ slant range vector from sensor to target (km; km/sec)
+            slant_range_sez (``ndarray``): 6x1 SEZ slant range vector from sensor to target (km; km/sec)
             noisy (``bool``, optional): whether measurements should include sensor noise. Defaults to ``False``.
 
         Returns:
             ``dict``: measurements made by the sensor
+
+            :``"azimuth_rad"``: (``float``): azimuth angle measurement (radians)
+            :``"elevation_rad"``: (``float``): elevation angle measurement (radians)
+            :``"range_km"``: (``float``): range measurement (km)
+            :``"range_rate_km_p_sec"``: (``float``): range rate measurement (km/sec)
         """
         raise NotImplementedError
 
@@ -100,16 +137,108 @@ class Sensor(metaclass=ABCMeta):
             - `2` is an angular measurement valid in [0, 2pi]
 
         Returns:
-            ``np.ndarray``: integer array defining angular measurements
+            ``ndarray``: integer array defining angular measurements
         """
         raise NotImplementedError
 
-    def buildSigmaObs(self, target_id, tgt_eci_state):
+    def collectObservations(
+        self, pointing_agent: EstimateAgent, background_agents: list[TargetAgent]
+    ) -> list[ObservationTuple]:
+        """Collect observations on all targets within the sensor's FOV.
+
+        Args:
+            pointing_agent (:class:`.EstimateAgent`): agent that sensor is pointing at
+            background_agents (``list``): list of possible `.TargetAgent` objects in FoV
+
+        Returns:
+            obs_list (``list``): list of `.ObservationTuple`
+        """
+        obs_list = []
+        # Check if sensor will slew to point in time
+        slant_range_sez = getSlantRangeVector(self.host.ecef_state, pointing_agent.eci_state)
+        if self.canSlew(slant_range_sez):
+            if self.calculate_field_of_view:
+                targets_in_fov = self.checkTargetsInView(slant_range_sez, background_agents)
+                obs_list = list(
+                    filter(  # pylint:disable=bad-builtin
+                        lambda x: x.observation,
+                        (
+                            self.makeNoisyObservation(
+                                tgt,
+                            )
+                            for tgt in targets_in_fov
+                        ),
+                    )
+                )  # pylint:disable=bad-builtin
+
+            else:
+                obs_list.append(self.makeNoisyObservation(pointing_agent))
+
+            self.boresight = slant_range_sez[:3] / norm(slant_range_sez[:3])
+            if obs_list:  # only update time_last_ob if obs are made
+                self.time_last_ob = self.host.time
+
+        return obs_list
+
+    def checkTargetsInView(
+        self, slant_range_sez: float, background_agents: list[TargetAgent]
+    ) -> list[TargetAgent]:
+        """Perform bulk FOV check on all RSOs.
+
+        Args:
+            slant_range_sez (``ndarray``): 3x1 slant range vector that sensor is pointing towards
+            background_agents (``list``): list of all `.TargetAgents` in scenario
+
+        Returns:
+            ``list``: list of all visible `.TargetAgents` in scenario
+        """
+        # filter out targets outside of FOV
+        agents_in_fov = list(
+            filter(  # pylint:disable=bad-builtin
+                lambda x: self.inFOV(
+                    slant_range_sez, getSlantRangeVector(self.host.ecef_state, x.eci_state)
+                ),
+                background_agents,
+            )
+        )  # pylint:disable=bad-builtin
+        return agents_in_fov
+
+    def inFOV(self, pointing_sez: ndarray, background_sez: ndarray) -> bool:
+        """Perform bulk FOV check on all RSOs.
+
+        Args:
+            pointing_sez (``ndarray``): 3x1 slant range vector that sensor is pointing towards
+            background_sez (``ndarray``): 3x1 slant range vector of possible agent in FoV
+
+        Returns:
+            ``bool``: whether or not a `.TargetAgent` object is within sensor FoV
+        """
+        if self.field_of_view.fov_shape == "conic":
+            angle = subtendedAngle(background_sez, pointing_sez)
+            return angle <= self.field_of_view.cone_angle / 2
+
+        if self.field_of_view.fov_shape == "rectangular":
+            pointing_azimuth = getAzimuth(pointing_sez)
+            pointing_elevation = getElevation(pointing_sez)
+
+            background_azimuth = getAzimuth(background_sez)
+            background_elevation = getElevation(background_sez)
+
+            azimuth_angle = abs(pointing_azimuth - background_azimuth)
+            elevation_angle = abs(pointing_elevation - background_elevation)
+            return (
+                azimuth_angle <= self.field_of_view.azimuth_angle / 2
+                and elevation_angle <= self.field_of_view.elevation_angle / 2
+            )
+
+        raise ValueError("Wrong Import for FoV Type")
+
+    def buildSigmaObs(self, target_id: int, tgt_eci_state: ndarray) -> Observation:
         """Calculate the measurement data for a sigma point observation.
 
         Args:
             target_id (``int``): simulation ID associated with current target.
-            tgt_eci_state (``np.ndarray``): 6x1 ECI state vector of the target (km; km/sec).
+            tgt_eci_state (``ndarray``): 6x1 ECI state vector of the target (km; km/sec).
 
         Returns:
             :class:`.Observation`: observation data object.
@@ -126,12 +255,14 @@ class Sensor(metaclass=ABCMeta):
             **self.getMeasurements(slant_range_sez),
         )
 
-    def makeObservation(self, tgt_id, tgt_eci_state, viz_cross_section, real_obs=False):
+    def makeObservation(
+        self, tgt_id: int, tgt_eci_state: ndarray, viz_cross_section: float, real_obs: bool = False
+    ) -> ObservationTuple:
         """Calculate the measurement data for a single observation.
 
         Args:
             tgt_id (``int``): simulation ID associated with current target
-            tgt_eci_state (``np.ndarray``): 6x1 ECI state vector of the target (km; km/sec)
+            tgt_eci_state (``ndarray``): 6x1 ECI state vector of the target (km; km/sec)
             viz_cross_section (``float``): visible cross-sectional area of the target (m^2)
             real_obs (``bool``, optional): whether observation should include noisy measurements. Defaults to ``False``.
 
@@ -156,7 +287,6 @@ class Sensor(metaclass=ABCMeta):
                     **self.getNoisyMeasurements(slant_range_sez),
                 )
                 self.boresight = slant_range_sez[:3] / norm(slant_range_sez[:3])
-                self.time_last_ob = self.host.time
                 self._current_target = tgt_id
 
             else:
@@ -168,7 +298,7 @@ class Sensor(metaclass=ABCMeta):
 
         return ObservationTuple(observation, self.host, self.angle_measurements)
 
-    def makeNoisyObservation(self, target_agent):
+    def makeNoisyObservation(self, target_agent) -> ObservationTuple:
         """Calculate the measurement data for a single observation with noisy measurements.
 
         Args:
@@ -200,27 +330,34 @@ class Sensor(metaclass=ABCMeta):
             real_obs=True,
         )
 
-    def getNoisyMeasurements(self, slant_range_sez):
+    def getNoisyMeasurements(self, slant_range_sez: ndarray) -> dict:
         """Return noisy measurements.
 
         Args:
-            slant_range_sez (``np.ndarray``): 6x1 SEZ slant range vector from sensor to target (km; km/sec)
+            slant_range_sez (``ndarray``): 6x1 SEZ slant range vector from sensor to target (km; km/sec)
 
         Returns:
             ``dict``: noisy measurements made by the sensor
+
+            :``"azimuth_rad"``: (``float``): azimuth angle measurement (radians)
+            :``"elevation_rad"``: (``float``): elevation angle measurement (radians)
+            :``"range_km"``: (``float``): range measurement (km)
+            :``"range_rate_km_p_sec"``: (``float``): range rate measurement (km/sec)
         """
         return self.getMeasurements(slant_range_sez, noisy=True)
 
-    def isVisible(self, tgt_eci_state, viz_cross_section, slant_range_sez):
+    def isVisible(
+        self, tgt_eci_state: ndarray, viz_cross_section: float, slant_range_sez: ndarray
+    ) -> bool:
         """Determine if the target is in view of the sensor.
 
         References:
             :cite:t:`vallado_2013_astro`, Sections 4.1 - 4.4 and 5 - 5.3.5
 
         Args:
-            tgt_eci_state (``np.ndarray``): 6x1 ECI state vector of the target agent
+            tgt_eci_state (``ndarray``): 6x1 ECI state vector of the target agent
             viz_cross_section (``float``): area of the target facing the sun (m^2)
-            slant_range_sez (``np.ndarray``): 6x1 SEZ slant range vector from sensor to target (km; km/sec)
+            slant_range_sez (``ndarray``): 6x1 SEZ slant range vector from sensor to target (km; km/sec)
 
         Returns:
             ``bool``: True if target is visible; False if target is not visible
@@ -240,47 +377,51 @@ class Sensor(metaclass=ABCMeta):
         azimuth = getAzimuth(slant_range_sez)
         elevation = getElevation(slant_range_sez)
 
-        ## [NOTE]: When `slant_range_sez` and `self.boresight` are colinear, lost precision creates a value __slightly__
-        ##          larger than 1.0 or smaller than -1.0. This results in domain errors with `numpy::arccos()`.
-        ##          This statement ensures to round to either 1 or -1 appropriately.
-        arg = dot(slant_range_sez[:3] / norm(slant_range_sez[:3]), self.boresight)
-        if abs(arg) - 1.0 > 0.0:
-            self.delta_boresight = arccos(fix(arg))
-        else:
-            self.delta_boresight = arccos(arg)
+        # Check if the elevation is within sensor bounds
+        if elevation <= self.el_mask[0] or elevation >= self.el_mask[1]:
+            return False
 
-        # Check if you are able to slew to the new target
-        if self.slew_rate * (self.host.time - self.time_last_ob) >= self.delta_boresight:
-            # Check if the elevation is within sensor bounds
-            if elevation <= self.el_mask[0] or elevation >= self.el_mask[1]:
-                return False
+        # [NOTE]: Azimuth check requires two versions:
+        #   - az_0 <= az_1 for normal situations
+        #   - az_0 > az_1 for when mask transits the 360deg/True North line
+        if self.az_mask[0] <= self.az_mask[1] and self.az_mask[0] <= azimuth <= self.az_mask[1]:
+            return True
 
-            # [NOTE]: Azimuth check requires two versions:
-            #   - az_0 <= az_1 for normal situations
-            #   - az_0 > az_1 for when mask transits the 360deg/True North line
-            if (
-                self.az_mask[0] <= self.az_mask[1]
-                and self.az_mask[0] <= azimuth <= self.az_mask[1]
-            ):
-                return True
-
-            if azimuth >= self.az_mask[0] or azimuth <= self.az_mask[1]:
-                return True
+        if azimuth >= self.az_mask[0] or azimuth <= self.az_mask[1]:
+            return True
 
         # Default: target satellite is not in view
         return False
 
+    def canSlew(self, slant_range_sez: ndarray) -> bool:
+        """Check if sensor can slew to target in the allotted time.
+
+        Args:
+            slant_range_sez (``ndarray``): slant range vector
+
+        Returns:
+            ``bool``: whether target can be slewed to in time
+        """
+        # [NOTE]: When `slant_range_sez` and `self.boresight` are colinear, lost precision creates a value
+        # __slightly__ larger than 1.0 or smaller than -1.0. This results in domain errors with `numpy::arccos()`.
+        # This statement ensures to round to either 1 or -1 appropriately.
+        arg = dot(slant_range_sez[:3] / norm(slant_range_sez[:3]), self.boresight)
+        self.delta_boresight = safeArccos(arg)
+        # Boolean if you are able to slew to the new target
+        # [TODO]: We are artificially increasing a sensor's slewing ability if it is not tasked at every timestep.
+        return self.slew_rate * (self.host.time - self.time_last_ob) >= self.delta_boresight
+
     @property
     def az_mask(self):
-        """``np.ndarray``: Returns the azimuth visibility mask."""
+        """``ndarray``: Returns the azimuth visibility mask."""
         return self._az_mask
 
     @az_mask.setter
-    def az_mask(self, new_az_mask):
+    def az_mask(self, new_az_mask: ndarray):
         """Ensure the azimuth mask is between [0, 2*pi].
 
         Args:
-            new_az_mask (``np.ndarray``): azimuth mask for a particular sensor
+            new_az_mask (``ndarray``): azimuth mask for a particular sensor
         """
         if isinstance(new_az_mask[0], (float, int)):
             if new_az_mask.all() >= 0 and new_az_mask.all() <= 2 * const.PI:
@@ -294,15 +435,15 @@ class Sensor(metaclass=ABCMeta):
 
     @property
     def el_mask(self):
-        """``np.ndarray``: Returns the elevation visibility mask."""
+        """``ndarray``: Returns the elevation visibility mask."""
         return self._el_mask
 
     @el_mask.setter
-    def el_mask(self, new_el_mask):
+    def el_mask(self, new_el_mask: ndarray):
         """Ensure the elevation mask is between [-pi, pi].
 
         Args:
-            new_el_mask (``np.ndarray``): elevation mask for a particular sensor
+            new_el_mask (``ndarray``): elevation mask for a particular sensor
         """
         if isinstance(new_el_mask[0], (float, int)):
             if new_el_mask.all() >= -const.PI and new_el_mask.all() <= const.PI:
@@ -316,15 +457,15 @@ class Sensor(metaclass=ABCMeta):
 
     @property
     def r_matrix(self):
-        """``np.ndarray``: Returns the measurement noise covariance matrix."""
+        """``ndarray``: Returns the measurement noise covariance matrix."""
         return self._r_matrix
 
     @r_matrix.setter
-    def r_matrix(self, new_specs):
+    def r_matrix(self, new_specs: ndarray):
         """Ensure the measurement noise matrix is a square matrix.
 
         Args:
-            new_specs (``np.ndarray``): measurement noise matrix for a particular sensor
+            new_specs (``ndarray``): measurement noise matrix for a particular sensor
         """
         if new_specs.shape[1] == 1 or new_specs.shape[0] == 1:
             self._r_matrix = diagflat(new_specs.ravel()) ** 2.0
@@ -348,7 +489,7 @@ class Sensor(metaclass=ABCMeta):
         """Assign host to an attribute, and sets other relevant properties accordingly.
 
         Args:
-            new_host (SensingAgent): containing `SensingAgent` object
+            new_host (:class:`.SensingAgent`): containing `SensingAgent` object
         """
         self._host = new_host
         self.time_last_ob = new_host.time
@@ -363,13 +504,11 @@ class Sensor(metaclass=ABCMeta):
         else:
             mid_el = self.el_mask[1] + (self.el_mask[0] - self.el_mask[1]) / 2.0
 
-        self.boresight = asarray(
-            [cos(mid_el) * cos(mid_az), cos(mid_el) * sin(mid_az), sin(mid_el)]
-        )
+        self.boresight = array([cos(mid_el) * cos(mid_az), cos(mid_el) * sin(mid_az), sin(mid_el)])
 
     @property
     def measurement_noise(self):
-        """``np.ndarray``: Returns randomly-generated measurement noise as v_k ~ N(0; r_matrix)."""
+        """``ndarray``: Returns randomly-generated measurement noise as v_k ~ N(0; r_matrix)."""
         return matmul(self._sqrt_noise_covar, random.randn(self._r_matrix.shape[0]))
 
     def getSensorData(self):
@@ -385,6 +524,7 @@ class Sensor(metaclass=ABCMeta):
             "aperture_area": self.aperture_area,
             "sensor_type": getTypeString(self),
             "exemplar": self.exemplar.tolist(),
+            "field_of_view": self.field_of_view,
         }
         if getTypeString(self.host) == "Spacecraft":
             output_dict["init_eci"] = self.host.truth_state.tolist()