+++ tests/data/events/test_sensor_addition.py
@@ -43,6 +43,8 @@ class TestSensorAdditionEventConfig(BaseTestCase):
                 "aperture_area": 0.8107319665559964,
                 "sensor_type": "Optical",
                 "exemplar": [0.0014320086173409336, 32500.0],
+                "field_of_view": {"fov_shape": "conic"},
+                "calculate_fov": False,
                 "lat": 0.1,
                 "lon": 1.1,
                 "alt": 1.0,
@@ -77,6 +79,8 @@ class TestSensorAdditionEventConfig(BaseTestCase):
                     -7.448488839044817,
                 ],
                 "exemplar": [1, 36000],
+                "field_of_view": {"fov_shape": "conic"},
+                "calculate_fov": False,
                 "host_type": "Spacecraft",
                 "station_keeping": {"routines": ["LEO"]},
             }
@@ -104,6 +108,8 @@ class TestSensorAdditionEventConfig(BaseTestCase):
                     "aperture_area": 0.8107319665559964,
                     "sensor_type": "Optical",
                     "exemplar": [0.0014320086173409336, 32500.0],
+                    "field_of_view": {"fov_shape": "conic"},
+                    "calculate_fov": False,
                     "host_type": "GroundFacility",
                 }
             )
@@ -130,6 +136,8 @@ class TestSensorAdditionEventConfig(BaseTestCase):
                     "aperture_area": 0.8107319665559964,
                     "sensor_type": "Optical",
                     "exemplar": [0.0014320086173409336, 32500.0],
+                    "field_of_view": {"fov_shape": "conic"},
+                    "calculate_fov": False,
                     "lat": 0.1,
                     "lon": 1.1,
                     "alt": 1.0,
@@ -168,6 +176,8 @@ class TestSensorAdditionEventConfig(BaseTestCase):
                     "sensor_type": "Optical",
                     "init_eci": bad_eci,
                     "exemplar": [1, 36000],
+                    "field_of_view": {"fov_shape": "conic"},
+                    "calculate_fov": False,
                     "host_type": "Spacecraft",
                     "station_keeping": {"routines": ["LEO"]},
                 }
@@ -203,6 +213,8 @@ class TestSensorAdditionEventConfig(BaseTestCase):
                         -7.448488839044817,
                     ],
                     "exemplar": [1, 36000],
+                    "field_of_view": {"fov_shape": "conic"},
+                    "calculate_fov": False,
                     "host_type": "Spacecraft",
                     "station_keeping": {"routines": ["LEO"]},
                 }
@@ -229,6 +241,8 @@ class TestSensorAdditionEventConfig(BaseTestCase):
                     "aperture_area": 0.8107319665559964,
                     "sensor_type": "Optical",
                     "exemplar": [0.0014320086173409336, 32500.0],
+                    "field_of_view": {"fov_shape": "conic"},
+                    "calculate_fov": False,
                     "lat": 0.1,
                     "lon": 1.1,
                     "alt": 1.0,
@@ -256,6 +270,8 @@ class TestSensorAdditionEventConfig(BaseTestCase):
                 "aperture_area": 0.8107319665559964,
                 "sensor_type": "Optical",
                 "exemplar": [0.0014320086173409336, 32500.0],
+                "field_of_view": {"fov_shape": "conic"},
+                "calculate_fov": False,
                 "lat": 0.1,
                 "lon": 1.1,
                 "alt": 1.0,
@@ -295,6 +311,10 @@ class TestSensorAdditionEvent(BaseTestCase):
                 "aperture_area": 0.8107319665559964,
                 "sensor_type": "Optical",
                 "exemplar": [0.0014320086173409336, 32500.0],
+                "field_of_view": {
+                    "fov_shape": "conic",
+                },
+                "calculate_fov": False,
                 "lat": 0.1,
                 "lon": 1.1,
                 "alt": 1.0,
@@ -333,6 +353,8 @@ class TestSensorAdditionEvent(BaseTestCase):
             sensor_type="Optical",
             exemplar_cross_section=0.0014320086173409336,
             exemplar_range=32500.0,
+            fov_shape="conic",
+            calculate_fov=False,
             station_keeping_json=dumps([]),
         )
         impulse_event.handleEvent(mocked_scenario)