+++ tests/estimation/test_adaptive_estimation.py
@@ -268,6 +268,8 @@ def getTestEarthSensor():
         diameter=27.0,
         efficiency=0.9,
         exemplar=array([1, 40744]),
+        field_of_view=1.0,
+        calculate_fov=False,
         power_tx=120000.0,
         frequency=10000000000.0,
         slew_rate=3.0000000000000004,